HTML addition
Adding HTML element, but struggling to get it to remain static on page.

I think I'm going to remove the drei Html element in the Page component and rather, in the Scene.jsx do some text handling for the html and what HTML (not drei Html, actual HTML) is visible.